Moved CfgGlasses to its own file
